+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/GroupRepository.php
@@ -270,30 +270,6 @@ class GroupRepository extends EntityRepository implements GroupRepositoryInterfa
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getOption($id, $collectionId = null, array $options = array())
-    {
-        // TODO: Implement getOption() method.
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOptionLabel($object, $dataLocale)
-    {
-        // TODO: Implement getOptionLabel() method.
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOptionId($object)
-    {
-        // TODO: Implement getOptionId() method.
-    }
-
-    /**
      * @return string
      */
     protected function getAlias()