+++ src/Pim/Bundle/TransformBundle/Denormalizer/Flat/ProductValue/ReferenceDataCollectionDenormalizer.php
@@ -23,6 +23,7 @@ class ReferenceDataCollectionDenormalizer extends AbstractValueDenormalizer
         ReferenceDataRepositoryResolver $repositoryResolver = null
     ) {
         parent::__construct($supportedTypes);
+
         $this->repositoryResolver = $repositoryResolver;
     }