+++ src/Pim/Component/ReferenceData/Model/AbstractReferenceData.php
@@ -69,6 +69,6 @@ abstract class AbstractReferenceData implements ReferenceDataInterface
      */
     public function __toString()
     {
-        return '[' . $this->code . ']';
+        return sprintf('[%s]', $this->code);
     }
 }